requirement.txt
langgraph
langchain
langchain-community
pydantic
pandas
scikit-learn
xgboost
lightgbm
mlflow
fastapi
uvicorn
matplotlib
seaborn
python-dotenv
openai 
azure-identity 
azure-storage-blob
azure-monitor-opentelemetry